Rename integrator hold in API and SCPI commands
The hold function freezes the internal state of the entire PID
controller and associated relock modules. Therefore "Integrator Hold" is
misleading. This commits renames the PID:IN#:OUT#:INT:HOLD SCPI command
to PID:IN#:OUT#:HOLD. The functions in the C API and python module are
also renamed to reflect the change.